+++ R/data-errors.R
@@ -1,0 +1,8 @@
+#' Errors
+#'
+#' These dataset contains KIS Open API errors
+#'
+#' @format data.frame
+#' @examples
+#' errors
+"errors"